dev-requirements.txt
Pillow==8.3.1
pygame==2.0.1
pygif==0.2
